+++ lib/src/models/condition.dart
@@ -17,16 +17,20 @@ abstract class Condition<T> extends ScreenModel<T> {
     T? desktop,
   }) : super(mobile: mobile, tablet: tablet, desktop: desktop);
 
-  const Condition.equals({T? value, this.conditional = Conditional.EQUALS})
-      : super(mobile: value, tablet: value, desktop: value);
+  const Condition.equals(T? value)
+      : conditional = Conditional.EQUALS,
+        super(mobile: value, tablet: value, desktop: value);
 
-  const Condition.largerThan(
-      {T? value, this.conditional = Conditional.LARGER_THAN})
-      : super(mobile: value, tablet: value, desktop: value);
+  const Condition.largerThan(T? value)
+      : conditional = Conditional.LARGER_THAN,
+        super(mobile: value, tablet: value, desktop: value);
 
-  const Condition.smallerThan(
-      {T? value, this.conditional = Conditional.SMALLER_THAN})
-      : super(mobile: value, tablet: value, desktop: value);
+  const Condition.smallerThan(T? value)
+      : conditional = Conditional.SMALLER_THAN,
+        super(mobile: value, tablet: value, desktop: value);
+
+  @override
+  String toString() => super.toString();
 }
 
 ///Has the type of conditionals for the [Condition]