+++ lib/src/theme/i_theme_data_rule.dart
@@ -9,5 +9,7 @@ import 'theme_rule.dart';
 ///
 /// {@endtemplate}
 mixin IThemeDataRule {
-  ThemeData getThemeByRule(ThemeRule rule);
+  ThemeData getThemeByRule([ThemeRule rule = ThemeRule.light]);
+
+  ThemeData getDefaultTheme() => getThemeByRule();
 }