+++ lib/src/utils/get_current_breakpoints.dart
@@ -1,3 +1,4 @@
+import 'package:argo/argo.dart';
 import 'package:argo/src/models/sb_value.dart';
 
 import '../models/screen_breakpoints.dart';
@@ -16,19 +17,12 @@ ScreenBreakpoints getCurrentBreakPoints({
   required ScreenBreakpoints global,
   ScreenBreakpoints? local,
 }) {
+  // the evaluation for SBValues is for MIN or MAX values.
   final bool isMin = local != null ? local.isMinSBValue : global.isMinSBValue;
 
-  final _mobile = isMin && global.mobile!.isMinType
-      ? global.mobile!
-      : defaultMinimumBreakPoints.mobile!;
-
-  final _tablet = isMin && global.tablet!.isMinType
-      ? global.mobile!
-      : defaultMinimumBreakPoints.tablet!;
-
-  final _desktop = isMin && global.desktop!.isMinType
-      ? global.desktop!
-      : defaultMinimumBreakPoints.desktop!;
+  final _mobile = _getDefaultOrValue(DeviceScreen.mobile, isMin, global);
+  final _tablet = _getDefaultOrValue(DeviceScreen.tablet, isMin, global);
+  final _desktop = _getDefaultOrValue(DeviceScreen.desktop, isMin, global);
 
   final defaultBreakPoints =
       isMin ? defaultMinimumBreakPoints : defaultMaximumBreakPoints;
@@ -38,6 +32,42 @@ ScreenBreakpoints getCurrentBreakPoints({
     tablet: _getDoubleValue(local?.tablet, _tablet),
     desktop: _getDoubleValue(local?.desktop, _desktop),
   );
+}
+
+SBValue _getDefaultOrValue(
+    DeviceScreen device, bool isMin, ScreenBreakpoints global) {
+  late SBValue _global;
+  late SBValue _defaultMin;
+  late SBValue _defaultMax;
+  switch (device) {
+    case DeviceScreen.mobile:
+      _global = global.mobile!;
+      _defaultMin = defaultMinimumBreakPoints.mobile!;
+      _defaultMax = defaultMaximumBreakPoints.mobile!;
+
+      break;
+    case DeviceScreen.tablet:
+      _global = global.tablet!;
+      _defaultMin = defaultMinimumBreakPoints.tablet!;
+      _defaultMax = defaultMaximumBreakPoints.tablet!;
+
+      break;
+    case DeviceScreen.desktop:
+      _global = global.desktop!;
+      _defaultMin = defaultMinimumBreakPoints.desktop!;
+      _defaultMax = defaultMaximumBreakPoints.desktop!;
+
+      break;
+  }
+  if (isMin == _global.isMinType) {
+    return _global;
+  } else if (!isMin && _global.isMinType) {
+    return _defaultMax;
+  } else if (isMin && !_global.isMinType) {
+    return _defaultMin;
+  }
+
+  return _global;
 }
 
 SBValue _getDoubleValue(