+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.4
+
+* Fix some problems with visibility widget
+* Add new widget to handle conditions
+* Add new example
+* Improve some classes and methods
+
+## 0.0.3
+
+* Add a new example reference
+* update some parts of theme rule
+
 ## 0.0.2
 
 * Add Documentation